Agenda
Made a start on the Agenda tiddler. It has a simple list all Action Point tiddlers which aren't tagged as "done" yet. There are icons which link to the relevant Action Point, Person, and/or Organisation. I tried using a <$reveal> widget to conditionally show the person-go and organisation-go links, but it didn't work.